Fixed devices not showing up on subscription page
The device column in the subscription page only showed ?? as there
where no device data stored in the DB, this is fixed in this commit.

Also added an additional column in the subscription page that shows
some more info about the registered action, currently only the play
position is displayed but it could be extended in the future.